New Fixes/Changes to PR #373 (#391)
* UWP Discover System

- Ability to fetch UWP programs and their logo.

* Fixed - able to obtain apps as LOCAL_SYSTEM

Before, it was only able to get "global" apps, but now it is able to get  from all User

* Fixed - Able to launch from the shortcuts now.

It was unreliable to search for Executable file. Therefore I decided to utilize `explorer shell:AppsFolder`.

This also required me to change how launching work in `winboat.ts`. by default `/app:program` cannot run `explorer.exe ...` method due to the argument are not allowed. instead we needed to pass to `,cmd:`

* Made the launching application more robust.

Before, I simply split the command by " ". Now we will have dedicated arg field. Figured it'll be useful in the future like giving custom launching arg to an application

* Context Menu

* editing, saving custom app

* clean up

* Update Apps.vue

* Update Apps.vue

* More Filter Option

* Fixed some inconsistency

* Added "Developer" filter for UWP app

* Update Apps.vue

* Resolves first Review

* npx prettier

* should resolve the review